% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{boxcox}
\alias{boxcox}
\title{Box-Cox transformation of variable}
\usage{
boxcox(x, lambda, na.rm = FALSE)
}
\arguments{
\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{

}
\description{
Box-Cox transformation of variable
}
\examples{

}
\references{
\url{http://onlinestatbook.com/2/transformations/box-cox.html}
}
