% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{skewness}
\alias{skewness}
\title{Skwness}
\usage{
skewness(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector containing the values whose skewness is to be computed.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The estimated skewness of \code{x}.
}
\description{
Skwness
}
\details{
An un-skewed distribution is one that is roughly symmetric. This means that the probability of falling on either side of the distribution's mean is roughly equal.

A right-skewed distribution has a large number of points on the left side of the distribution (smaller values) than on the right side (larger values).

If the predictor distribution is roughly symmetric, the skewness values will be close to zero. As the distribution becomes more right skewed, the skewness statistic becomes larger. Similarly, as the distribution becomes more left skewed, the value becomes negative.
}
\examples{
x <- rnorm(100)
skewness(x)
}
\references{
447 Applied Predictive Modeling (page 31)
}
